Basic Saving & Loading
Can save and load player health

+++ Assets/Scripts/DamageReport.cs
@@ -2,5 +2,5 @@ public class DamageReport
 {
     public EntityController causedBy;
     public EntityController target;
-    public double damage;
+    public float damage;
 }
